Created the New-DSCPullServer function
This script tests whether the machine can access the internet and then tests to see if there is not already a Pull server installed. If all is ok it will then download and install all the required DSC modules and features and then proceed to create and test the DSC pull server on the localhost. This implementation is using http: only,